reorganization commit
