+++ src/main/java/pl/sda/librarymanagementapp/model/book/BookRepository.java
@@ -1,7 +1,7 @@
 package pl.sda.librarymanagementapp.model.book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pl.sda.librarymanagementapp.domain.Book;
+import pl.sda.librarymanagementapp.domain.book.Book;
 
 public interface BookRepository extends JpaRepository<Book,Long> {
 }
